Removed TextFamilyFieldType in favor of StringFieldType
